4.14.17
- Added more compiler/linker flags for debug/release builds
- Cleaned up all level 3 warnings!
- Testing enforcing level 3 warnings to be treated as errors!
- Not building with level 4 warnings, as libraries we depend on haven't cleaned up warnings that far, and they bleed through at W4. Also, we need to keep unreferenced formal parameters in headers for more clear documentation.
  - This solves issue #82

+++ src/Modules/Editor/UI/Editor_Interface.cpp
@@ -88,7 +88,7 @@ void Editor_Interface::tick(const float& deltaTime)
 	ImGui::ShowDemoWindow(&show_demo_window);
 
 	// Prepare the docking regions
-	const auto dockspace_size = ImVec2(m_renderSize.x / 5.0F, m_renderSize.y);
+	const auto dockspace_size = ImVec2((float)m_renderSize.x / 5.0F, (float)m_renderSize.y);
 	const auto window_flags = ImGuiWindowFlags_NoTitleBar |
 		ImGuiWindowFlags_MenuBar |
 		ImGuiWindowFlags_NoMove |
